@@ -1,45 +1,24 @@
-import React, { useRef, useEffect } from 'react';
-import { useLocation, Switch } from 'react-router-dom';
-import AppRoute from './utils/AppRoute';
-import ScrollReveal from './utils/ScrollReveal';
-import ReactGA from 'react-ga';
+import "./App.css";
+import { ContactUs } from "./components/ContactUs";
+import { Header } from "./components/Header";
+import { Mission } from "./components/Mission";
+import { Services } from "./components/Services";
+import { TeamMembers } from "./components/TeamMembers";
+import { WhatWeDo } from "./components/WhatWeDo";
 
-// Layouts
-import LayoutDefault from './layouts/LayoutDefault';
-
-// Views 
-import Home from './views/Home';
-
-// Initialize Google Analytics
-ReactGA.initialize(process.env.REACT_APP_GA_CODE);
-
-const trackPage = page => {
-  ReactGA.set({ page });
-  ReactGA.pageview(page);
-};
-
-const App = () => {
-
-  const childRef = useRef();
-  let location = useLocation();
-
-  useEffect(() => {
-    const page = location.pathname;
-    document.body.classList.add('is-loaded')
-    childRef.current.init();
-    trackPage(page);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [location]);
-
-  return (
-    <ScrollReveal
-      ref={childRef}
-      children={() => (
-        <Switch>
-          <AppRoute exact path="/" component={Home} layout={LayoutDefault} />
-        </Switch>
-      )} />
-  );
+function App() {
+	return (
+		<div id="home" className="App bg-white min-h-screen">
+			<div className="max-w-7xl mx-auto grid gap-10 md:gap-14 p-8">
+				<Header />
+				<WhatWeDo />
+				<Mission />
+				<Services />
+				<TeamMembers />
+				<ContactUs />
+			</div>
+		</div>
+	);
 }
 
 export default App;
